requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
dnspython>=2.4.0
